added to original inventory file
